Remove non-constructor code
